requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic>=2.5.3
pytest>=8.0.0
httpx>=0.26.0
python-jose>=3.3.0
slowapi>=0.1.8
redis>=5.0.1
starlette~=0.45.3